+++ sky/clouds/service_catalog/constants.py
@@ -1,6 +1,6 @@
 """Constants used for service catalog."""
-import os
-
 HOSTED_CATALOG_DIR_URL = 'https://raw.githubusercontent.com/skypilot-org/skypilot-catalog/master/catalogs'  # pylint: disable=line-too-long
 CATALOG_SCHEMA_VERSION = 'v5'
 CATALOG_DIR = '~/.sky/catalogs'
+ALL_CLOUDS = ('aws', 'azure', 'gcp', 'ibm', 'lambda', 'scp', 'oci',
+              'kubernetes', 'runpod', 'vsphere', 'cudo', 'fluidstack')
