+++ sky/execution.py
@@ -20,7 +20,7 @@ from sky import sky_logging
 from sky import spot
 from sky import task as task_lib
 from sky.backends import backend_utils
-from sky.clouds import service_catalog
+from sky.clouds.service_catalog import common as service_catalog_common
 from sky.skylet import constants
 from sky.usage import usage_lib
 from sky.utils import common_utils
@@ -677,7 +677,7 @@ def spot_launch(
             'retry_until_up': retry_until_up,
             'remote_user_config_path': remote_user_config_path,
             'modified_catalogs':
-                service_catalog.get_modified_catalog_file_mounts(),
+                service_catalog_common.get_modified_catalog_file_mounts(),
             **controller_utils.shared_controller_vars_to_fill(
                 'spot',
                 remote_user_config_path=remote_user_config_path,
